Add @using(..) helper to pass objects to PhantomJS.
This hack has the same limitations as when passing variables from PhantomJS
to Lotte - it can only work on simple types and cannot serialize functions,
regular expressions, etc.

It is still better than having nothing or having to write the code yourself
each time a variable from a different environment is needed.