Add a dummy HTTP server.
This server will be used for IPC with PhantomJS. Events, such as test
case start, pass, fail, etc. will be communicated over HTTP (not yet
implemented).

The server is enabled by default, but can be disabled either using
'--no-server' on the command line or 'server = false' in a Lottefile.

The default listening port is 3838 which should be unprivileged.
The default bind address is localhost, 127.0.0.1.